Fix laggy input issue on intersect page
